♻️ Extracted URL parsing utilities from App.js to url-parsers.js
Extracted three pure functions to improve code organization and testability:
- parseOfferQueryString() - Parse offer preview query parameters
- parsePreviewQueryString() - Parse preview mode query parameters
- parsePortalLinkPath() - Parse portal link paths to page/query

Reduced App.js by 244 lines (1046 → 802 lines)
All tests passing (256/257)
